+++ plotnine/guides/guide_colorbar.py
@@ -151,8 +151,9 @@ class guide_colorbar(guide):
 
         # 1.45 makes the default colourbar wider than the
         # legend entry boxes.
-        width = (self.barwidth or _d('legend_key_width')) * 1.45
-        height = (self.barheight or _d('legend_key_height')) * 1.45
+        width = (self.barwidth or _d('legend_key_width') or 16) * 1.45
+        height = (self.barheight or _d('legend_key_height') or 16) * 1.45
+
         height *= 5
         length = height